+++ src/App.tsx
@@ -2,8 +2,9 @@ import React, { useState } from 'react';
 import ChickenTable from './components/ChickenTable/ChickenTable';
 import ProductsFilter from './components/Products/ProductsFilter';
 import { ChickenData } from './types/ChickenData';
-import DemoPage from './component/DemoPage'
+import DemoPage from "./components/DemoPage";
 import './App.css';
+import DetailSideTable from "./components/Details/DetailSideTable";
 
 const App: React.FC = () => {
   const [filteredData, setFilteredData] = useState<ChickenData[]>([]);
@@ -12,6 +13,7 @@ const App: React.FC = () => {
     <div>
       <ProductsFilter onFilter={setFilteredData} />
       <ChickenTable filteredData={filteredData} />
+      <DetailSideTable />
     </div>
   );
 }
