+++ src/components/Details/DetailSideTable.tsx
@@ -8,23 +8,29 @@ import {
   tableWrapper,
   title,
   wrapper,
-} from "./Sidebar.css";
-import { ChickenData } from "../../types/ChickenData";
-import { FC } from "react";
+} from "./DetailSideTable.css";
+import { useDispatch, useSelector } from "react-redux";
+import { AppDispatch, RootState } from "../../store";
 
-type TSidebarProps = {
-  selectedItems: ChickenData[];
-}
+const DetailSideTable = () => {
+  const dispatch = useDispatch<AppDispatch>();
+  const detailOnSelectedItems = useSelector(
+    (state: RootState) => state.details
+  );
 
-const Sidebar: FC<TSidebarProps> = ({ selectedItems }) => {
+  const handleOnCloseButton = () => {
+    dispatch();
+  };
 
-  const handleOnClose = () => {};
+  if (!detailOnSelectedItems.selectedData[0]) {
+    return <></>;
+  }
 
   return (
     <div className={wrapper}>
       <div className={header}>
         <div className={title}> 🍗 내가 가장 원하는 닭가슴살 제품은 ? 🍗 </div>
-        <FiX className={closeButton} onClick={handleOnClose} />
+        <FiX className={closeButton} onChange={handleOnCloseButton} />
       </div>
       <table className={tableWrapper}>
         <thead className={tableHeader}>
@@ -45,7 +51,7 @@ const Sidebar: FC<TSidebarProps> = ({ selectedItems }) => {
           </tr>
         </thead>
         <tbody className={tableBody}>
-          {selectedItems.map((item) => (
+          {detailOnSelectedItems.selectedData.map((item) => (
             <tr key={item.id} className={tableRow}>
               <td>{item.id}</td>
               <td>{item.brand ? item.brand : "-"}</td>
@@ -68,4 +74,4 @@ const Sidebar: FC<TSidebarProps> = ({ selectedItems }) => {
   );
 };
 
-export default Sidebar;
+export default DetailSideTable;
